Maven Domain: use replace text
